chore(greenhouse): simplifies the store
